+++ src/Tracker/index.tsx
@@ -1,15 +1,15 @@
 import { h, Component, ComponentType } from 'preact'
-import { BrowserClient, EventHint, Hub, Severity } from '@sentry/browser'
 import { cleanFalsy, wrapArray } from '~utils/array'
-import WoopraTracker from './safeWoopra'
 import { map as mapObject } from '~utils/object'
-import { isOnfidoHostname } from '~utils/string'
+import * as execeptionTracking from '~core/ExceptionHandler'
+import * as Woopra from '~core/Woopra'
 import { sendAnalyticsEvent } from './onfidoTracker'
 import { integratorTrackedEvents } from './trackerData'
 import { v4 as uuidv4 } from 'uuid'
 const Cookies = require('js-cookie')
 
 import type { TrackScreenCallback, WithTrackingProps } from '~types/hocs'
+import type { EventHint } from '~core/ExceptionHandler'
 import type {
   LegacyTrackedEventNames,
   UserAnalyticsEventNames,
@@ -18,77 +18,23 @@ import type {
 
 let shouldSendEvents = false
 
-const client = window.location.hostname
-const sdk_version = process.env.SDK_VERSION
-let sentryClient: BrowserClient | undefined
-let sentryHub: Hub | undefined
-let woopra: WoopraTracker = null
-
-const setUp = (): void => {
-  woopra = new WoopraTracker('onfidojssdkwoopra')
-
-  woopra.init()
-
-  // configure tracker
-  woopra.config({
-    domain: process.env.WOOPRA_DOMAIN,
-    cookie_name: 'onfido-js-sdk-woopra',
-    cookie_domain: location.hostname,
-    referer: location.href,
-  })
-
-  // Do not overwrite the woopra client if we are in the cross device client.
-  // This is so we can track the original page where the user opened the SDK.
-  woopra.identify(
-    client.match(/^(id|id-dev)\.onfido\.com$/)
-      ? { sdk_version }
-      : { sdk_version, client }
-  )
+const setUp = () => {
+  Woopra.install()
+  shouldSendEvents = true
 }
 
 const uninstall = (): void => {
-  if (sentryClient) {
-    sentryClient.close(2000).then(() => {
-      sentryClient = undefined
-      sentryHub = undefined
-      process.exit()
-    })
-  }
+  execeptionTracking.uninstall()
   uninstallWoopra()
 }
 
 const uninstallWoopra = (): void => {
-  woopra && woopra.dispose()
+  Woopra.uninstall()
   shouldSendEvents = false
 }
 
 const install = (): void => {
-  sentryClient = new BrowserClient({
-    dsn: 'https://6e3dc0335efc49889187ec90288a84fd@sentry.io/109946',
-    environment: process.env.NODE_ENV,
-    release: sdk_version,
-    debug: true,
-    // TODO: Make sure the whitelisting works as expected
-    whitelistUrls: [/onfido[A-z.]*\.min.js/g],
-    beforeBreadcrumb: (crumb) => {
-      const isOnfidoXhr =
-        crumb.category === 'xhr' && isOnfidoHostname(crumb.data?.url)
-
-      const isOnfidoClick =
-        crumb.category === 'ui.click' &&
-        crumb.message?.includes('.onfido-sdk-ui')
-
-      const shouldReturnCrumb = isOnfidoXhr || isOnfidoClick
-
-      return shouldReturnCrumb ? crumb : null
-    },
-    // @TODO: verify these mismatched options
-    // autoBreadcrumbs: { console: false },
-    // shouldSendCallback: () => process.env.PRODUCTION_BUILD,
-  })
-  sentryHub = new Hub(sentryClient)
-  sentryHub.addBreadcrumb({ level: Severity.Info })
-
+  execeptionTracking.install()
   shouldSendEvents = true
 }
 
@@ -129,7 +75,7 @@ const sendEvent = (
 
   if (shouldSendEvents) {
     const formattedProperties = formatProperties(properties)
-    woopra && woopra.track(eventName, formattedProperties)
+    Woopra.track(eventName, formattedProperties)
     sendAnalyticsEvent(eventName, formattedProperties)
   }
 }
@@ -183,29 +129,14 @@ const trackComponent = <P extends WithTrackingProps>(
   }
 
 const trackException = (message: string, extra?: EventHint): void => {
-  sentryHub?.captureException(new Error(message), extra)
-}
-
-const setWoopraCookie = (cookie: string): void => {
-  if (!woopra) {
-    return
-  }
-
-  const cookie_name = woopra.config('cookie_name')
-  const cookie_expire = woopra.config('cookie_expire')
-  const cookie_path = woopra.config('cookie_path')
-  const cookie_domain = woopra.config('cookie_domain')
-  woopra.docCookies.setItem(
-    cookie_name,
-    cookie,
-    cookie_expire,
-    cookie_path,
-    cookie_domain
+  execeptionTracking?.captureException(
+    new Error(message),
+    extra as Record<string, unknown>
   )
-  woopra.cookie = cookie
 }
 
-const getWoopraCookie = (): Optional<string> => (woopra ? woopra.cookie : null)
+const setWoopraCookie = Woopra.setCookie
+const getWoopraCookie = Woopra.getCookie
 
 const cookieAttributes = {
   name: 'onfido-web-sdk-analytics',