+++ test/mock-server/local/common.js
@@ -11,6 +11,24 @@
     }
   }
 
+  window.request = function(method, url, sessionId, payload, statusCode) {
+    var xhr = new XMLHttpRequest();
+    xhr.open(method, url, false);
+    xhr.setRequestHeader("X-Session-Id", sessionId)
+
+    if (payload === "") {
+      xhr.setRequestHeader("Content-Type", "text/plain")
+    } else {
+      xhr.setRequestHeader("Content-Type", "application/json")
+    }
+
+    if (statusCode) {
+      xhr.setRequestHeader("X-Status", statusCode)
+    }
+
+    xhr.send(payload);
+  }
+
   var parameters = new URLSearchParams(location.search);
   if (parameters.has('link_id')) {
     window.onfido = Onfido.init({