+++ test/mock-server/routers.ts
@@ -1,6 +1,6 @@
 /* eslint-disable no-duplicate-imports */
-import { Router, Status } from './deps.ts'
-import type { FormDataBody } from './deps.ts'
+import type { FormDataBody, RouterContext } from './deps.ts'
+import { deepMerge, Router, Status } from './deps.ts'
 
 import { generateToken } from './jwt.ts'
 import responses from './responses.ts'
@@ -192,14 +192,95 @@ telephonyRouter.post('/v1/cross_device_sms', (context) => {
   context.response.body = responses.telephony.v1.cross_device_sms
 })
 
+type MockCode =
+  | 'sdkConfiguration'
+  | 'consents'
+  | 'workflowRun'
+  | 'workflowRunComplete'
+
+type MockValue = {
+  status: number
+  value: any
+  patch: boolean
+}
+
+const allowedCodes: MockCode[] = [
+  'sdkConfiguration',
+  'workflowRun',
+  'workflowRunComplete',
+]
+const mockResponses: Record<string, Record<MockCode, MockValue>> = {}
+
+const mockRouter = new Router({ prefix: '/mock' })
+
+const mockHandler = async (context: any) => {
+  const sessionId = context.request.headers.get('X-Session-Id')
+  const code = context.params.code as MockCode
+  const status = context.request.headers.get('X-Status')
+
+  if (!sessionId || allowedCodes.indexOf(code) === -1) {
+    context.response.status = Status.BadRequest
+  }
+
+  mockResponses[sessionId as string] = mockResponses[sessionId as string] || {}
+  mockResponses[sessionId as string][code] = {
+    patch: 'patch' === context.request.method.toLowerCase(),
+    value: await context.request.body().value,
+    status: status ? status : Status.OK,
+  }
+
+  context.response.status = Status.OK
+}
+
+mockRouter.patch('/:code', mockHandler)
+mockRouter.put('/:code', mockHandler)
+
+const getMock = (sessionId: any, code: MockCode): MockValue | undefined => {
+  if (
+    sessionId &&
+    mockResponses[sessionId] &&
+    code in mockResponses[sessionId]
+  ) {
+    return mockResponses[sessionId][code]
+  }
+  return
+}
+
+const mockPayload = (mock: MockValue, defaultPayload: any) => {
+  if (mock.patch) {
+    return deepMerge(defaultPayload, mock.value)
+  }
+
+  return mock.value
+}
+
+const sendMock = (context: any, code: MockCode, defaultPayload: any) => {
+  const sessionId = context.request.headers.get('X-Session-Id')
+
+  context.response.body = defaultPayload
+  context.response.status = Status.OK
+
+  if (!sessionId) {
+    return
+  }
+
+  const mock = getMock(sessionId, code)
+
+  if (!mock) {
+    return
+  }
+
+  context.response.body = mockPayload(mock, defaultPayload)
+  context.response.status = mock.status
+}
+
 const apiRouter = new Router({ prefix: '/api' })
 apiRouter
   .get('/ping', (context) => {
     context.response.body = { message: 'pong' }
   })
   .post('/v3.3/sdk/configurations', async (context) => {
-    context.response.body = responses.api.v3.sdk_configurations
-    context.response.status = Status.OK
+    sendMock(context, 'sdkConfiguration', responses.api.v3.sdk_configurations)
   })
   .patch('/v3.3/applicants/:id/location', async (context) => {
     context.response.status = Status.OK
@@ -208,8 +289,7 @@ apiRouter
     context.response.status = Status.OK
   })
   .get('/v3.3/applicants/:id/consents', async (context) => {
-    context.response.body = responses.api.v3.applicant_consents
-    context.response.status = Status.OK
+    sendMock(context, 'consents', responses.api.v3.applicant_consents)
   })
   .get(
     '/v3.3/report_types/proof_of_address/supported_countries',
@@ -276,5 +356,11 @@ apiRouter
     await sleep(500)
     context.response.body = responses.api.v4.documents
   })
+  .get('/v4/workflow_runs/:id', (context) => {
+    sendMock(context, 'workflowRun', {})
+  })
+  .post('/v4/workflow_runs/:id/complete', (context) => {
+    sendMock(context, 'workflowRunComplete', {})
+  })
 
-export { apiRouter, telephonyRouter, tokenFactoryRouter }
+export { mockRouter, apiRouter, telephonyRouter, tokenFactoryRouter }