+++ test/mock-server/server.ts
@@ -1,7 +1,12 @@
 import { Application, oakCors } from './deps.ts'
 
 import { frontendMiddleware, loggerMiddleware } from './middlewares.ts'
-import { apiRouter, telephonyRouter, tokenFactoryRouter } from './routers.ts'
+import {
+  mockRouter,
+  apiRouter,
+  telephonyRouter,
+  tokenFactoryRouter,
+} from './routers.ts'
 
 const app = new Application()
 
@@ -9,6 +14,9 @@ app.use(loggerMiddleware)
 app.use(oakCors())
 
 /* Back-end routes */
+
+app.use(mockRouter.routes())
+app.use(mockRouter.allowedMethods())
 app.use(apiRouter.routes())
 app.use(apiRouter.allowedMethods())
 app.use(tokenFactoryRouter.routes())