+++ test/webtest/src/test/java/com/onfido/qa/websdk/test/ProofOfAddressIT.java
@@ -4,16 +4,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.onfido.qa.websdk.PoADocumentType;
 import com.onfido.qa.websdk.model.CompleteData;
 import com.onfido.qa.websdk.model.DocumentOption;
-import com.onfido.qa.websdk.page.Complete;
-import com.onfido.qa.websdk.page.CrossDeviceClientIntro;
-import com.onfido.qa.websdk.page.CrossDeviceClientSuccess;
-import com.onfido.qa.websdk.page.CrossDeviceMobileConnected;
-import com.onfido.qa.websdk.page.CrossDeviceSubmit;
-import com.onfido.qa.websdk.page.DocumentUpload;
-import com.onfido.qa.websdk.page.IdDocumentSelector;
-import com.onfido.qa.websdk.page.ImageQualityGuide;
-import com.onfido.qa.websdk.page.PoADocumentSelection;
-import com.onfido.qa.websdk.page.PoAIntro;
+import com.onfido.qa.websdk.page.*;
 import com.onfido.qa.websdk.sdk.PoAStep;
 import com.onfido.qa.websdk.sdk.Raw;
 import org.testng.SkipException;
@@ -33,7 +24,6 @@ import static com.onfido.qa.websdk.PoADocumentType.UTILITY_BILL;
 import static com.onfido.qa.websdk.PoADocumentType.values;
 import static com.onfido.qa.websdk.UploadDocument.NATIONAL_IDENTITY_CARD_PDF;
 import static com.onfido.qa.websdk.UploadDocument.PASSPORT_JPG;
-import static com.onfido.qa.websdk.UploadDocument.UK_DRIVING_LICENCE_PNG;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @SuppressWarnings("StaticCollection")
@@ -197,8 +187,9 @@ public class ProofOfAddressIT extends WebSdkIT {
                 .select(BANK_BUILDING_SOCIETY_STATEMENT)
                 .clickContinue()
                 .upload(NATIONAL_IDENTITY_CARD_PDF)
-                .clickConfirmButton(IdDocumentSelector.class)
-                .select(PASSPORT, DocumentUpload.class)
+                .clickConfirmButton(RestrictedDocumentSelection.class)
+                .selectSupportedCountry()
+                .selectDocument(PASSPORT, DocumentUpload.class)
                 .clickUploadButton(ImageQualityGuide.class)
                 .upload(PASSPORT_JPG)
                 .clickConfirmButton(Complete.class);
@@ -222,8 +213,9 @@ public class ProofOfAddressIT extends WebSdkIT {
     public void testPassportDocFollowedByPoa() throws JsonProcessingException {
         onfido().withSteps("document", "poa", "complete")
                 .withOnComplete(new Raw("(data) => {window.onCompleteData = data}"))
-                .init(IdDocumentSelector.class)
-                .select(PASSPORT, DocumentUpload.class)
+                .init(RestrictedDocumentSelection.class)
+                .selectSupportedCountry()
+                .selectDocument(PASSPORT, DocumentUpload.class)
                 .clickUploadButton(ImageQualityGuide.class)
                 .upload(PASSPORT_JPG)
                 .clickConfirmButton(PoAIntro.class)